fanvue-matrix-sync/requirements.txt
matrix-nio[e2e]
aiohttp
../fanvue_common
